config/requirements/production.txt
-r base.txt

# uWSGI==2.0.22
psycopg2==2.9.8
whitenoise==6.6.0
gunicorn==21.2.0
# uvicorn==0.29.0
dj-database-url==2.1.0
# whitenoise[brotli]>=6.0,<7.0
# psycopg[c]; sys_platform == "linux"
# psycopg[binary]; sys_platform != "linux"
# psycopg2-binary
django-storages
boto3
